pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytorchlightning_example"
version = "1.0.0"
description = "Federated Learning with PyTorch Lightning and Flower (Quickstart Example)"
license = "Apache-2.0"
dependencies = [
    "torch==2.6.0",
    "hydra-core==1.3.2",
    "omegaconf>=2.3.0",
    "flwr==1.18.0",
    "fedlab==1.3.0",
    "matplotlib==3.10.3",
    "numpy==2.0.1",
    "flwr[simulation]>=1.18.0",
    "flwr-datasets[vision]>=0.5.0",
    "pytorch-lightning==2.4.0",
    "torchvision==0.21.0",
    "wandb"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "src.server_app:app"
clientapp = "src.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
max-epochs = 1
config-name = "mnist_cnn"


[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10